Move UnitEnum to online-workflow project Add getUnit method in OnlineUtils
