+++ online-workflow/src/main/java/eu/itesla_project/online/UnitEnum.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package eu.itesla_project.online.rest.model;
+package eu.itesla_project.online;
 
 import com.fasterxml.jackson.annotation.JsonValue;
